app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
motor==3.3.2
email-validator==2.1.0
httpx==0.26.0
python-dotenv==1.0.0

cloudinary==1.36.0
google-auth==2.27.0
requests==2.31.0
python-multipart==0.0.20